+++ plugins/balloontoolbar/plugin.js
@@ -163,6 +163,8 @@
 	CKEDITOR.ui.balloonToolbar.prototype.deleteItem = function( name ) {
 		if ( this._items[ name ] ) {
 			delete this._items[ name ];
+			// Redraw balloon toolbar when item is removed.
+			this._view.renderItems( this._items );
 		}
 	};
 
@@ -418,7 +420,8 @@
 			if ( !selection ) {
 				selection = this.editor.getSelection();
 
-				// Shrink the selection so that we're ensured innermost elements are available.
+				// Shrink the selection so that we're ensured innermost elements are available, so that path for
+				// selection like `foo [<em>bar</em>] baz` also contains `em` element.
 				CKEDITOR.tools.array.forEach( selection.getRanges(), function( range ) {
 					range.shrink( CKEDITOR.SHRINK_ELEMENT, true );
 				} );
@@ -466,6 +469,14 @@
 
 			// Match element selectors.
 			if ( path ) {
+				// First check the outermost element (if any was selected), since the selection got shrinked
+				// it would be otherwise skipped (#1274).
+				var selectedElem = selection.getSelectedElement();
+
+				if ( selectedElem && !selectedElem.isReadOnly() ) {
+					matchEachContext( this._contexts, elementsMatcher, selectedElem );
+				}
+
 				for ( var i = 0; i < path.elements.length; i++ ) {
 					var curElement = path.elements[ i ];
 					// Skip non-editable elements (e.g. widget internal structure).
@@ -591,6 +602,8 @@
 				CKEDITOR.ui.balloonPanel.prototype.build.call( this );
 				this.parts.panel.addClass( 'cke_balloontoolbar' );
 				this.parts.title.remove();
+				// Following workaround is needed until #1370 is resolved.
+				this.deregisterFocusable( this.parts.close );
 				this.parts.close.remove();
 			};
 
@@ -601,21 +614,16 @@
 				var editable = this.editor.editable();
 				this._detachListeners();
 
-				this._listeners.push( this.editor.on( 'resize', function() {
+				function attachListener() {
 					this.attach( this._pointedElement, {
 						focusElement: false
 					} );
-				}, this ) );
-				this._listeners.push( editable.attachListener( editable.getDocument(), 'scroll', function() {
-					this.attach( this._pointedElement, {
-						focusElement: false
-					} );
-				}, this ) );
-				this._listeners.push( CKEDITOR.document.getWindow().on( 'resize', function() {
-					this.attach( this._pointedElement, {
-						focusElement: false
-					} );
-				}, this ) );
+				}
+
+				this._listeners.push( this.editor.on( 'change', attachListener, this ) );
+				this._listeners.push( this.editor.on( 'resize', attachListener, this ) );
+				this._listeners.push( CKEDITOR.document.getWindow().on( 'resize', attachListener, this ) );
+				this._listeners.push( editable.attachListener( editable.getDocument(), 'scroll', attachListener, this ) );
 
 				CKEDITOR.ui.balloonPanel.prototype.show.call( this );
 			};
@@ -664,6 +672,7 @@
 			};
 
 			CKEDITOR.ui.balloonToolbarView.prototype.destroy = function() {
+				this._deregisterItemFocusables();
 				CKEDITOR.ui.balloonPanel.prototype.destroy.call( this );
 				this._detachListeners();
 			};
@@ -679,6 +688,9 @@
 					keys = CKEDITOR.tools.objectKeys( items ),
 					groupStarted = false;
 
+				// When we rerender toolbar we want to clear focusable in case of removing some items.
+				this._deregisterItemFocusables();
+
 				CKEDITOR.tools.array.forEach( keys, function( itemKey ) {
 
 					// If next element to render is richCombo and we have already opened group we have to close it.
@@ -690,7 +702,6 @@
 						groupStarted = true;
 						output.push( '<span class="cke_toolgroup">' );
 					}
-
 					// Now we can render element.
 					items[ itemKey ].render( this.editor, output );
 				}, this );
@@ -704,7 +715,8 @@
 				this.parts.content.unselectable();
 				CKEDITOR.tools.array.forEach( this.parts.content.find( 'a' ).toArray(), function( element ) {
 					element.setAttribute( 'draggable', 'false' );
-				} );
+					this.registerFocusable( element );
+				}, this );
 			};
 
 			/**
@@ -722,6 +734,22 @@
 				this._pointedElement = element;
 
 				CKEDITOR.ui.balloonPanel.prototype.attach.call( this, element, options );
+			};
+
+			/**
+			 * Method deregisters {@link #focusables} registered by UI items, like buttons.
+			 *
+			 * @private
+			 * @since 4.8.1
+			 * @member CKEDITOR.ui.balloonToolbarView
+			 */
+			CKEDITOR.ui.balloonToolbarView.prototype._deregisterItemFocusables = function() {
+				var focusables = this.focusables;
+				for ( var id in focusables ) {
+					if ( this.parts.content.contains( focusables[ id ] ) ) {
+						this.deregisterFocusable( focusables[ id ] );
+					}
+				}
 			};
 		}
 	} );